Add proccntmax, corecnt, corecntHwloc, getVendorID
proccnt:
We can not use CPUID.1:EBX[23:16] to figure out the number
of logical processors, because it "represents the maximum
number of addressable IDs (initial APIC ID) that can be assigned
to logical processors in a physical package. The value may not
be the same as the number of logical processors that are present
in the hardware of a physical package."
software.intel.com/en-us/articles/
    intel-64-architecture-processor-topology-enumeration
software.intel.com/sites/default/files/63/1a/
    Kuo_CpuTopology_rc1.rh1.final.pdf
This leaves us with the SYSTEM_INFO structure for the Windows
version and glibc's sysconf extension for the the Linux version.

proccntmax:
This new function returns the maximum number of logical
processors per physical package based on CPUID.1:EBX[23:16].
See above.

corecnt:
This new function returns the number of processor cores.
So far, there is only a version for Linux that parses
/proc/cpuinfo.

corecntHwloc:
This new function is an alternative to the function corecnt
that uses and depends on the Portable Hardware Locality (hwloc)
software package.

getVendorID:
This new function returns the CPU's manufacturer ID string
based on CPUID.